+++ app/Http/Controllers/CategoriasController.php
@@ -8,20 +8,24 @@ use App\Models\Categoria;
 class CategoriasController extends Controller
 {
     public function index(){
+        $this->authorize('view',Categoria::class);
         $categorias = Categoria::paginate(10);
         return view('categorias.index', compact('categorias'));
     }
 
     public function create(){
+        $this->authorize('create',Categoria::class);
         return view('categorias.create');
     }
 
     public function edit($id){
+        $this->authorize('update',Categoria::class);
         $categoria = Categoria::findOrFail($id);
         return view('categorias.edit', compact('categoria'));
     }
 
     public function store(Request $request){
+        $this->authorize('create',Categoria::class);
         $request->validate([
             'nombre' => 'required',
             'activo' => 'required|boolean'
@@ -32,6 +36,7 @@ class CategoriasController extends Controller
 
     public function update(Request $request, $id){
         //dd($request->all(),$id);
+        $this->authorize('update',Categoria::class);
         $request->validate([
             'nombre' => 'required',
             'activo' => 'required|boolean'
@@ -43,11 +48,13 @@ class CategoriasController extends Controller
     }
 
     public function show($id){
+        $this->authorize('view',Categoria::class);
         $categoria = Categoria::findOrFail($id);
         return view('categorias.show', compact('categoria'));
     }
 
     public function desactivate($id){
+        $this->authorize('delete',Categoria::class);
         $categoria = Categoria::findOrFail($id);
         $categoria->activo = false;
         $categoria->save();
@@ -55,6 +62,7 @@ class CategoriasController extends Controller
     }
 
     public function destroy($id){
+        $this->authorize('delete',Categoria::class);
         $categoria = Categoria::findOrFail($id);
         $categoria->delete();
         return redirect()->route('categorias.index')->with('success', 'Categoría eliminada exitosamente.');
