+++ app/Http/Controllers/FacultadesController.php
@@ -10,24 +10,25 @@ use Illuminate\Support\Facades\Auth;
 class FacultadesController extends Controller
 {
     public function index(){
-        $this->authorize('view',Auth::user());
+        $this->authorize('view',Facultad::class);
         $facultades = Facultad::paginate(6);
         return view('facultades.index', compact('facultades'));
     }
 
     public function create(){
-        $this->authorize('create',Auth::user());
+        $this->authorize('create',Facultad::class);
         return view('facultades.create');
     }
 
     public function edit($id){
-        $this->authorize('update',Auth::user());
+        $this->authorize('update',Facultad::class);
         $facultad = Facultad::findOrFail($id);
         return view('facultades.edit', compact('facultad'));
     }
 
     public function store(Request $request){
         //dd($request->all());
+        $this->authorize('create',Facultad::class);
         $request->validate([
             'nombre' => 'required',
             'activo' => 'required|boolean'
@@ -38,6 +39,7 @@ class FacultadesController extends Controller
     }
 
     public function update(Request $request, $id){
+        $this->authorize('update',Facultad::class);
         $request->validate([
             'nombre' => 'required',
             'activo' => 'required|boolean'
@@ -49,11 +51,13 @@ class FacultadesController extends Controller
     }
 
     public function show($id){
+        $this->authorize('view',Facultad::class);
         $facultad = Facultad::findOrFail($id);
         return view('facultades.show', compact('facultad'));
     }
 
     public function desactivate($id){
+        $this->authorize('delete',Facultad::class);
         $facultad = Facultad::findOrFail($id);
         $facultad->activo = false;
         $facultad->save();
@@ -61,6 +65,7 @@ class FacultadesController extends Controller
     }
 
     public function destroy($id){
+        $this->authorize('delete',Facultad::class);    
         $facultad = Facultad::findOrFail($id);
         $facultad->delete();
         return redirect()->route('facultades.index')->with('success', 'Facultad eliminada exitosamente.');
