+++ app/Http/Controllers/PermisosController.php
@@ -8,20 +8,24 @@ use Spatie\Permission\Models\Permission;
 class PermisosController extends Controller
 {
     public function index(){
+        $this->authorize('view',Permission::class);
         $permisos = Permission::paginate(6);
         return view('permisos.index', compact('permisos'));
     }
 
     public function create(){
+        $this->authorize('create',Permission::class);
         return view('permisos.create');
     }
 
     public function edit($id){
+        $this->authorize('update',Permission::class);
         $permiso = Permission::findOrFail($id);
         return view('permisos.edit', compact('permiso'));
     }
 
     public function store(Request $request){
+        $this->authorize('create',Permission::class);
         $request->validate([
             'name' => 'required|unique:permissions,name',
         ]);
@@ -31,6 +35,7 @@ class PermisosController extends Controller
     }
 
     public function update(Request $request, $id){
+        $this->authorize('update',Permission::class);
         $request->validate([
             'name' => 'required|unique:permissions,name,' . $id,
         ]);
@@ -41,17 +46,20 @@ class PermisosController extends Controller
     }
 
     public function show($id){
+        $this->authorize('view',Permission::class);
         $permiso = Permission::findOrFail($id);
         return view('permisos.show', compact('permiso'));
     }
 
     public function destroy($id){
+        $this->authorize('delete',Permission::class);
         $permiso = Permission::findOrFail($id);
         $permiso->delete();
         return redirect()->route('permisos.index')->with('success', 'Permiso eliminado exitosamente.');
     }
     
     public function desactivate($id){
+        $this->authorize('delete',Permission::class);
         $permiso = Permission::findOrFail($id);
         $permiso->activo = false;
         $permiso->save();
